+++ src/main/kotlin/de/sepdetect/sepdetect/controller/PatientController.kt
@@ -8,6 +8,7 @@ import de.sepdetect.sepdetect.repository.PatientRepository
 import de.sepdetect.sepdetect.service.impl.UserService
 import de.sepdetect.sepdetect.util.HttpResponseStatus
 import de.sepdetect.sepdetect.util.JsonViews
+import org.springframework.security.authentication.InsufficientAuthenticationException
 import org.springframework.web.bind.annotation.*
 import java.lang.IllegalArgumentException
 import java.time.LocalDate
@@ -15,7 +16,9 @@ import java.time.format.DateTimeFormatter
 import java.util.*
 import javax.persistence.EntityNotFoundException
 
-
+/**
+ * Kontroller zum erstellen, manipulieren und löschen von Patienten.
+ */
 @CrossOrigin
 @RestController
 class PatientController constructor(
@@ -23,6 +26,10 @@ class PatientController constructor(
         private val organizationRepository: OrganizationRepository,
         private val userService: UserService) {
 
+    /**
+     * Gibt eine Liste mit allen aktiven Patienten zurück, auf die der Benutzter zugreifen darf. Ist der Benutzer ein Admin
+     * werden alle aktiven Patienten zurückgegeben.
+     */
     @JsonView(JsonViews.PatientsOnly::class)
     @GetMapping("patients")
     fun getPatients(): List<Patient> {
@@ -30,12 +37,25 @@ class PatientController constructor(
         return if (user.role == UserRole.ADMIN) patientRepository.findAllByActive(true) else patientRepository.findAllByActiveAndOrganizationIn(true, userService.getCurrentUser().organization)
     }
 
+    /**
+     * Endpunkt gibt den Patienten und alle Score-Werte mit der entsprechenden Patienten-ID als JSON-Objekt zurück.
+     * Wirt einen Fehler wenn der Benutzer nicht auf den Patienten zugreifen darf.
+     */
     @JsonView(JsonViews.FullPatient::class)
     @GetMapping("patient/get/{id}")
     fun getPatient(@PathVariable id: Long): Patient {
+        val pat = patientRepository.findPatientByPersonId(id).orElseThrow { throw EntityNotFoundException("Patient not found!") }
+        val user = userService.getCurrentUser()
+        if (user.role != UserRole.ADMIN && !user.organization.any { it.id == pat.organization?.id })
+            throw EntityNotFoundException("Access denied!")
+
         return patientRepository.findPatientByPersonId(id).orElseThrow { throw EntityNotFoundException("Patient not found!") }
     }
 
+    /**
+     * Ermöglicht es einen neuen Patienten zu erstellen. Zum Erstellen eines neuen Patienten muss dem Endpunkt ein Patienten-Objekt übergeben werden.
+     * Der Patient muss zwingend einer Organisation zugeordnet sein, sonst wird ein Fehler geworfen.
+     */
     @JsonView(JsonViews.FullPatient::class)
     @PostMapping("patient/add")
     fun addPatient(@RequestBody patient: Patient): Patient {
@@ -50,7 +70,7 @@ class PatientController constructor(
     }
 
     /**
-     *
+     *  Ermöglicht des bearbeiten von Patienten.
      */
     @JsonView(JsonViews.FullPatient::class)
     @PutMapping("patient/edit")
@@ -75,15 +95,22 @@ class PatientController constructor(
     }
 
     /**
-     *
+     *  Löscht den angebenden Patienten aus der Datenbank. Alle SOFA-Verläufe werden gemeinsam mit dem Patienten gelöscht.
+     *  Der ausführende Benutzter muss die Rolle Admin inne haben.
      */
     @DeleteMapping("patient/delete/{id}")
     fun deletePatient(@PathVariable id: Long): HttpResponseStatus {
+        if (userService.getCurrentUser().role != UserRole.ADMIN)
+            throw InsufficientAuthenticationException("Not Permitted!")
+
         val pat = patientRepository.findPatientByPersonId(id).orElseThrow { throw EntityNotFoundException("Patient not found!") }
         patientRepository.delete(pat)
         return HttpResponseStatus("OK", "Patient deleted (id: $id)")
     }
 
+    /**
+     * Aktiviert oder deaktiviert einen Patienten.
+     */
     @GetMapping("patient/active/{patientID}")
     fun togglePatient(@PathVariable patientID: Long, @RequestParam("active") active: Optional<Boolean>): HttpResponseStatus {
         if (!active.isPresent)
@@ -100,6 +127,10 @@ class PatientController constructor(
         return HttpResponseStatus("OK", "Set Patient to ${if (active.get()) "active" else "inactive"}")
     }
 
+    /**
+     * Sucht Patienten anhand von bestimmten Kriterien. Alle Parameter sind optional. Wird kein Parameter angeben wird ein
+     * Fehler ausgegeben.
+     */
     @JsonView(JsonViews.PatientsOnly::class)
     @GetMapping("patient/search")
     fun searchPatients(@RequestParam("lastname") lastname: Optional<String>, @RequestParam("surname") surname: Optional<String>, @RequestParam("birthday") birthday: Optional<String>, @RequestParam("gender") gender: Optional<Char>): List<Patient> {
@@ -108,8 +139,8 @@ class PatientController constructor(
 
         val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
 
-        val lastName = if (lastname.isPresent && lastname.get().isNotEmpty()) String(Base64.getDecoder().decode(lastname.get())) else null
-        val surName = if (surname.isPresent && surname.get().isNotEmpty()) String(Base64.getDecoder().decode(surname.get())) else null
+        val lastName = if (lastname.isPresent && lastname.get().isNotEmpty()) String(Base64.getDecoder().decode(lastname.get())).toLowerCase() else null
+        val surName = if (surname.isPresent && surname.get().isNotEmpty()) String(Base64.getDecoder().decode(surname.get())).toLowerCase() else null
         val birthday = if (birthday.isPresent && birthday.get().isNotEmpty()) LocalDate.parse(String(Base64.getDecoder().decode(birthday.get())), formatter) else null
         val gender = if (gender.isPresent) gender.get() else null
